+++ RELEASES.md
@@ -6,6 +6,9 @@ This new release adds support for sparse cost matrices and a new lazy EMD solver
 
 #### New features 
 - Add lazy EMD solver with on-the-fly distance computation from coordinates (PR #788)
+- Add Warmstart feature to the EMD solver for existing potentials (PR #793)
+- Add Warmstart potentials feature to the EMD solver for lazy and sparse solver (PR #795)
+- Faster init and result retrieval for EMD solver (PR #796)
 - Migrate backend from deprecated `scipy.sparse.coo_matrix` to modern `scipy.sparse.coo_array` (PR #782)
 - Geomloss function now handles both scalar and slice indices for i and j (PR #785)
 - Add support for sparse cost matrices in EMD solver (PR #778, Issue #397)
@@ -21,6 +24,8 @@ This new release adds support for sparse cost matrices and a new lazy EMD solver
 - Fix openmp flags on macOS (PR #789)
 - Clean documentation (PR #787)
 - Fix code coverage (PR #791)
+- Fix test of the version of jax in `ot.backend` (PR #794)
+- Reverting the openmp fix on macOS (PR #789) for macOS (PR #797)
 
 
 ## 0.9.6.post1